Tidy up use of CPLRecode, including fixing a probable memory leak. CPLRecode() may call CPLError. Do we wish to override the error handling ? In dumpPPM() skip the vrc_magic36 case if VRC36_SKIP is defined ! VRGetUInt(): implicit type promotion promotes to *signed* int;   the sign messes up the left-shift, so be explicit that we promote to unsigned.

+++ frmts/vrc/VRCutils.cpp
@@ -30,10 +30,10 @@ unsigned int VRGetUInt(const void* base, unsigned int byteOffset )
 {
     const unsigned char* buf =
         static_cast<const unsigned char*>(base)+byteOffset;
-    int vv = buf[0];
-    vv |= (buf[1] << 8U);
-    vv |= (buf[2] << 16U);
-    vv |= (buf[3] << 24U);
+    unsigned int vv = static_cast<unsigned int>(buf[0]);
+    vv |= static_cast<unsigned int>(buf[1]) << 8U;
+    vv |= static_cast<unsigned int>(buf[2]) << 16U;
+    vv |= static_cast<unsigned int>(buf[3]) << 24U;
 
     return(static_cast<unsigned int>(vv));
 }
